requirements_linux.txt
asgiref==3.8.1
black==24.4.2
boto3==1.35.9
botocore==1.35.9
cached-property==1.5.2
certifi==2024.6.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
dj-url-filter==0.4.4
Django==5.1.3
django-compiler==0.45
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
# django-rest==0.8.7
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-queryfields==1.1.0
drf-yasg==1.21.8
et-xmlfile==1.1.0
greenlet==3.1.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==2.0.0
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.10
pycparser==2.22
pyparsing==3.1.4
pypdf==4.3.1
PyPDFForm==1.4.33
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python_minifier==2.11.2
pytz==2024.2
#pywin32==306
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
s3transfer==0.10.2
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
waitress==3.0.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwings==0.32.0
xlwt==1.3.0
